Loyalty Lookup Validation
Added validation and messages to the loyalty lookup page. The user will now receive errors when trying to search with no value entered, and also if they try to search a user id that doesn't exist. I changed the type of the search bar from 'search' to 'number' to eliminate users from trying to search with words.